[arccore,alina] Use 'int' instead of 'ptrdiff_t' in MPI Alina Solver.
The 64bit version will be added in a future version.

+++ arccore/src/alina/arccore/alina/AlinaLib.h
@@ -115,8 +115,8 @@ class ARCCORE_ALINA_EXPORT AlinaLib
   // Create distributed solver.
   static AlinaDistributedSolver* solver_mpi_create(MPI_Comm comm,
                                                    ptrdiff_t n,
-                                                   const ptrdiff_t* ptr,
-                                                   const ptrdiff_t* col,
+                                                   const int* ptr,
+                                                   const int* col,
                                                    const double* val,
                                                    int n_def_vec,
                                                    AlinaDefVecFunction def_vec_func,
